add typescript configuration
