+++ tsconfig.json
@@ -1,0 +1,18 @@
+{
+    "compilerOptions": {
+      "module": "es6",
+      "target": "es5",
+      "jsx": "react",
+      "jsxFactory": "h",
+      "strict": true,
+      "isolatedModules": true,
+      "esModuleInterop": true,
+      "allowSyntheticDefaultImports": true,
+      "noUnusedLocals": false,
+      "allowJs": true,
+      "outDir": "dist",
+      "lib": ["dom", "es6"],
+      "moduleResolution": "node",
+    },
+    "exclude": ["node_modules", "public", ".cache", "example", "dist", "**/*.test.js"]
+}
